+++ test/finder_test.rb
@@ -6,9 +6,9 @@ class FinderTest < TestHelper
   def test_initialize_from_module
     finder = BrokenLinkFinder.new sort: :link, max_threads: 10
 
-    assert_equal({}, finder.broken_links)
-    assert_equal({}, finder.ignored_links)
-    assert_equal 0, finder.total_links_crawled
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_empty finder.crawl_stats
     assert_equal 10, finder.max_threads
     assert_equal :link, finder.sort
     assert_equal Set, finder.instance_variable_get(:@all_broken_links).class
@@ -21,9 +21,9 @@ class FinderTest < TestHelper
   def test_initialize
     finder = Finder.new
 
-    assert_equal({}, finder.broken_links)
-    assert_equal({}, finder.ignored_links)
-    assert_equal 0, finder.total_links_crawled
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_empty finder.crawl_stats
     assert_equal 100, finder.max_threads
     assert_equal :page, finder.sort
     assert_equal Set, finder.instance_variable_get(:@all_broken_links).class
@@ -37,15 +37,18 @@ class FinderTest < TestHelper
     assert_equal 10, finder.max_threads
   end
 
-  def test_clear_links
+  def test_reset_crawl
     finder = Finder.new
     finder.instance_variable_set :@broken_links, name: 'foo'
     finder.instance_variable_set :@ignored_links, name: 'bar'
-    finder.clear_links
+    finder.instance_variable_set :@crawl_stats, { url: '/blah' }
+    finder.instance_variable_set :@all_broken_links, ['/blah']
+    finder.instance_variable_set :@all_intact_links, ['/blah']
+    finder.reset_crawl
 
     assert_empty finder.broken_links
     assert_empty finder.ignored_links
-    assert_equal 0, finder.total_links_crawled
+    assert_empty finder.crawl_stats
     assert_empty finder.instance_variable_get(:@all_broken_links)
     assert_empty finder.instance_variable_get(:@all_intact_links)
   end
@@ -66,7 +69,12 @@ class FinderTest < TestHelper
                      'tel:+13174562564'
                    ]
                  }, finder.ignored_links)
-    assert_equal 8, finder.total_links_crawled
+
+    assert_equal 'http://mock-server.com/', finder.crawl_stats[:url]
+    assert_equal ['http://mock-server.com/'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 8, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__sort_by_link
@@ -89,25 +97,38 @@ class FinderTest < TestHelper
                      'http://mock-server.com/'
                    ]
                  }, finder.ignored_links)
-    assert_equal 8, finder.total_links_crawled
+
+    assert_equal 'http://mock-server.com/', finder.crawl_stats[:url]
+    assert_equal ['http://mock-server.com/'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 8, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__no_broken_links
     finder = Finder.new
     refute finder.crawl_url('http://mock-server.com/location')
 
-    assert_equal({}, finder.broken_links)
-    assert_equal({}, finder.ignored_links)
-    assert_equal 2, finder.total_links_crawled
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_equal 'http://mock-server.com/location', finder.crawl_stats[:url]
+    assert_equal ['http://mock-server.com/location'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 2, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__no_broken_links__sort_by_link
     finder = Finder.new sort: :link
     refute finder.crawl_url('http://mock-server.com/location')
 
-    assert_equal({}, finder.broken_links)
-    assert_equal({}, finder.ignored_links)
-    assert_equal 2, finder.total_links_crawled
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_equal 'http://mock-server.com/location', finder.crawl_stats[:url]
+    assert_equal ['http://mock-server.com/location'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 2, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__links_page
@@ -135,9 +156,14 @@ class FinderTest < TestHelper
         'https://thisdomaindoesnotexist-thouthou.com/nonexistentimage.png'
       ]
     }
+
     assert_equal expected, finder.broken_links
     assert_empty finder.ignored_links
-    assert_equal 15, finder.total_links_crawled
+    assert_equal 'https://example.co.uk/links.html', finder.crawl_stats[:url]
+    assert_equal ['https://example.co.uk/links.html'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 15, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__links_page__sort_by_link
@@ -165,7 +191,12 @@ class FinderTest < TestHelper
     }
     assert_equal expected, finder.broken_links
     assert_empty finder.ignored_links
-    assert_equal 15, finder.total_links_crawled
+
+    assert_equal 'https://example.co.uk/links.html', finder.crawl_stats[:url]
+    assert_equal ['https://example.co.uk/links.html'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 15, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_crawl_url__invalid
@@ -175,22 +206,15 @@ class FinderTest < TestHelper
     flunk
   rescue RuntimeError => e
     assert_equal 'Invalid or broken URL: https://server-error.com', e.message
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_empty finder.crawl_stats
   end
 
   def test_crawl_site
     finder = Finder.new
-    has_broken_links, crawled_pages = finder.crawl_site 'http://mock-server.com/'
+    assert finder.crawl_site 'http://mock-server.com/'
 
-    assert has_broken_links
-    assert_equal([
-                   'http://mock-server.com/',
-                   'http://mock-server.com/contact',
-                   'http://mock-server.com/location',
-                   'http://mock-server.com/about',
-                   'http://mock-server.com/not_found',
-                   'http://mock-server.com/location?q=hello',
-                   'http://mock-server.com/about?q=world'
-                 ], crawled_pages)
     assert_equal({
                    'http://mock-server.com/' => [
                      'https://doesnt-exist.com',
@@ -217,26 +241,29 @@ class FinderTest < TestHelper
                      'ftp://websiteaddress.com'
                    ]
                  }, finder.ignored_links)
-    assert_equal 17, finder.total_links_crawled
+
+    assert_equal 'http://mock-server.com/', finder.crawl_stats[:url]
+    assert_equal([
+      'http://mock-server.com/',
+      'http://mock-server.com/contact',
+      'http://mock-server.com/location',
+      'http://mock-server.com/about',
+      'http://mock-server.com/not_found',
+      'http://mock-server.com/location?q=hello',
+      'http://mock-server.com/about?q=world'
+    ], finder.crawl_stats[:pages_crawled])
+    assert_equal 7, finder.crawl_stats[:num_pages]
+    assert_equal 17, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
 
     # Check it can be run multiple times consecutively without error.
-    finder.crawl_site 'http://mock-server.com/'
+    assert finder.crawl_site 'http://mock-server.com/'
   end
 
   def test_crawl_site__sort_by_link
     finder = Finder.new sort: :link
-    has_broken_links, crawled_pages = finder.crawl_site 'http://mock-server.com/'
+    assert finder.crawl_site 'http://mock-server.com/'
 
-    assert has_broken_links
-    assert_equal([
-                   'http://mock-server.com/',
-                   'http://mock-server.com/contact',
-                   'http://mock-server.com/location',
-                   'http://mock-server.com/about',
-                   'http://mock-server.com/not_found',
-                   'http://mock-server.com/location?q=hello',
-                   'http://mock-server.com/about?q=world'
-                 ], crawled_pages)
     assert_equal({
                    '#doesntexist' => [
                      'http://mock-server.com/contact'
@@ -257,11 +284,23 @@ class FinderTest < TestHelper
                    'mailto:youraddress@yourmailserver.com' => ['http://mock-server.com/'],
                    'tel:+13174562564' => ['http://mock-server.com/']
                  }, finder.ignored_links)
-    assert_equal 17, finder.total_links_crawled
+
+    assert_equal 'http://mock-server.com/', finder.crawl_stats[:url]
+    assert_equal([
+      'http://mock-server.com/',
+      'http://mock-server.com/contact',
+      'http://mock-server.com/location',
+      'http://mock-server.com/about',
+      'http://mock-server.com/not_found',
+      'http://mock-server.com/location?q=hello',
+      'http://mock-server.com/about?q=world'
+    ], finder.crawl_stats[:pages_crawled])
+    assert_equal 7, finder.crawl_stats[:num_pages]
+    assert_equal 17, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
 
     # Check it can be run multiple times consecutively without error.
-    has_broken_links, = finder.crawl_site 'http://mock-server.com/'
-    assert has_broken_links
+    assert finder.crawl_site 'http://mock-server.com/'
   end
 
   def test_crawl_site__invalid
@@ -271,6 +310,9 @@ class FinderTest < TestHelper
     flunk
   rescue RuntimeError => e
     assert_equal 'Invalid or broken URL: https://server-error.com', e.message
+    assert_empty finder.broken_links
+    assert_empty finder.ignored_links
+    assert_empty finder.crawl_stats
   end
 
   def test_retry_mechanism
@@ -279,6 +321,11 @@ class FinderTest < TestHelper
 
     assert_empty finder.broken_links
     assert_empty finder.ignored_links
+    assert_equal 'http://www.retry.com', finder.crawl_stats[:url]
+    assert_equal ['http://www.retry.com'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 1, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 
   def test_retry_mechanism__sort_by_link
@@ -287,5 +334,10 @@ class FinderTest < TestHelper
 
     assert_empty finder.broken_links
     assert_empty finder.ignored_links
+    assert_equal 'http://www.retry.com', finder.crawl_stats[:url]
+    assert_equal ['http://www.retry.com'], finder.crawl_stats[:pages_crawled]
+    assert_equal 1, finder.crawl_stats[:num_pages]
+    assert_equal 1, finder.crawl_stats[:num_links]
+    assert finder.crawl_stats[:duration] > 0.0
   end
 end